Refactor, add clap struct
