+++ Cargo.toml
@@ -1,10 +1,12 @@
 [package]
 name = "agency"
+author = ["Stephen Hara <stephen.hara@stephenhara.com>"]
 version = "0.1.0"
+about = "CLI helping you manage SSH agents when they get gnarly"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-clap = "4.2.7"
+clap = { version = "4.2.7", features = ["derive"]}
 inquire = "0.6.2"
